requirements.txt
aio-pika
PyYAML
aiopg
git+git://github.com/NQN-Discord/rabbit_helper.git